+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## Unreleased
+
+## 1.0.0 - 2020-11-03
+
+### Fixed
+
+-   Double the Donation settings have been reviewed (#6)
+-   DtD form setting is now hidden until enabled (#7)
+-   Form setting removes the feature when disabled (#8)